pyproject.toml
[tool.poetry]
name = "cgcalc"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
bs4 = "^0.0.2"
google-api-python-client = "^2.119.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
